# internal/isms/api/openapi.yaml
openapi: '3.1.0'
info:
  title: isms.sh API
  description: |
    AI-powered Information Security Management System.
    Git-native document management with ISO 27001, NIS2, and SOC 2 support.

    ## Authentication

    Three authentication methods (checked in order):

    1. **API tokens** - `Authorization: Bearer isms_xxx` - for CLI and AI agents
    2. **JWT session tokens** - `Authorization: Bearer <jwt>` - from login endpoint
    3. **Cloudflare Zero Trust** - JWT verification via `Cf-Access-Jwt-Assertion` header

    ## Multi-tenancy

    Organization context is resolved from subdomain, custom domain, or headers:
    - `X-Organization-UUID` header (preferred)
    - `X-Organization` header (slug)
    - `?org=` query parameter
  version: '1.0.0'
  contact:
    name: isms.sh
    url: https://isms.sh
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api/v1
    description: API v1
security:
  - bearerAuth: []
  - apiKey: []
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
                otp_code: { type: string, description: TOTP code if OTP is enabled }
                org_slug: { type: string, description: Organization slug to log into }
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string, description: JWT session token }
                  user: { $ref: '#/components/schemas/User' }
                  organization: { $ref: '#/components/schemas/Organization' }
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/verify-email:
    post:
      tags: [Auth]
      summary: Verify email with code
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email: { type: string, format: email }
                code: { type: string }
      responses:
        '200':
          description: Email verified
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /auth/passkey/login/begin:
    post:
      tags: [Auth]
      summary: Begin passkey login ceremony
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200':
          description: WebAuthn credential request options
  /auth/passkey/login/complete:
    post:
      tags: [Auth]
      summary: Complete passkey login ceremony
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: WebAuthn assertion response
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user: { $ref: '#/components/schemas/User' }
...
  /auth/oidc/authorize:
    get:
      tags: [Auth]
      summary: Begin OIDC authorization flow
      security: []
      parameters:
        - name: provider
          in: query
          required: true
          schema: { type: string }
      responses:
        '302':
          description: Redirect to OIDC provider
  /auth/oidc/callback:
    get:
      tags: [Auth]
      summary: OIDC authorization callback
      security: []
      parameters:
        - name: code
          in: query
          schema: { type: string }
        - name: state
          in: query
          schema: { type: string }
      responses:
        '302':
          description: Redirect to app with session token

  # ── Users ───────────────────────────────────────────────────────────────
...
  /organizations:
    post:
      tags: [Users]
      summary: Create a new organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug]
              properties:
                name: { type: string }
                slug: { type: string, pattern: '^[a-z0-9-]+$' }
      responses:
        '201':
          description: Organization created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Organization' }

  # ── Config ──────────────────────────────────────────────────────────────
...
  /reviews:
    get:
      tags: [Reviews & Comments]
      summary: List all reviews
      responses:
        '200':
          description: List of reviews
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Review' }
    post:
      tags: [Reviews & Comments]
      summary: Create a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [document_id]
              properties:
                document_id: { type: string }
                document_type: { type: string }
                title: { type: string }
                assignees:
                  type: array
                  items: { type: string, format: email }
      responses:
        '201':
          description: Review created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Review' }
...
  /comments:
    post:
      tags: [Reviews & Comments]
      summary: Add a comment to a document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [document_id, body]
              properties:
                document_id: { type: string }
                review_id: { type: integer }
                body: { type: string }
                section: { type: string }
                paragraph_index: { type: integer }
                quote: { type: string }
                parent_id: { type: integer, description: For threaded replies }
      responses:
        '201':
          description: Comment added
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Comment' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT session token from login
    apiKey:
      type: http
      scheme: bearer
      description: "Personal API key (isms_ prefix)"
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        message: { type: string }
    User:
      type: object
      properties:
        id: { type: integer }
        email: { type: string, format: email }
        name: { type: string }
        otp_verified: { type: boolean }
        email_verified: { type: boolean }
        active: { type: boolean }
        created_at: { type: string, format: date-time }
        last_seen: { type: string, format: date-time }
...
    Organization:
      type: object
      properties:
        uuid: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        domain: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    Review:
      type: object
      properties:
        id: { type: integer }
        document_id: { type: string }
        document_type: { type: string }
        title: { type: string }
        version: { type: string }
        commit_hash: { type: string }
        requested_by: { type: string }
        status: { type: string, enum: [open, in_review, approved, changes_requested, closed] }
        comment_count: { type: integer }
        open_comments: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Comment:
      type: object
      properties:
        id: { type: integer }
        review_id: { type: integer }
        document_id: { type: string }
        author: { type: string }
        body: { type: string }
        section: { type: string }
        paragraph_index: { type: integer }
        quote: { type: string }
        parent_id: { type: integer }
        status: { type: string, enum: [open, resolved] }
        resolved_by: { type: string }
        resolved_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
...